style: renamed adminProfile
